elevator_reporter/requirements.txt
mysql-connector-python
pandas
openpyxl
python-dotenv